+++ projects/ui-lib/src/lib/services/index.ts
@@ -1,2 +1,4 @@
 export * from './mapper';
 export * from './screen-sizer';
+export * from './theme';
+